moved IconLinks into its own file, cleaned up ProjectModal.js
